eopkg: Add unified eopkg py3/nuitka'd package
**Summary**
This package would replace:
- The existing pure-python3 eopkg4 draft PR from Fabio
- python-eopkg (from Joey), which contains only the eopkg4 pure python
  libraries
- eopkg4-bin (from Hans), which is the currently in-testing method for
  getting python3-based eopkg4 code + PackageKit code onto Staff systems.

Hence, this package contains both the python3 library code, the pure
python3 executables, the nuitka'd eopkg.bin standalone executable,
and the nuitka'd eopkgBackend.bin PK backend.

Like ypkg, this package would enable like-for-like comparison between
the pure .py3 scripts and nuitka'd standalone binariy archive, and would
control whether the py2 or the py3 versions are enabled on user systems
via symlinks as a packaging operation.

The resulting benefit to Solus is that we don't have to bump two or three
different packages each time we make a change, but have it all
consolidated in one place.

Note that the eopkg py2 implementation we currently use is called pisi,
hence the suggestion by Joey to call this eopkg.

Signed-off-by: Rune Morling <ermo@serpentos.com>